Add support for Argon2id password hashes
This change provides an alternative to plaintext password verification.
The plaintext method is not resistant to timing attacks and allows the
password to be extracted from /data/admind.passwd. With this change, if
/data/passwd.argon2id exists, passwords are verified against it and
plaintext verification is disabled, even if any errors occur. Also,
logging has been added.

The time cost is 2, the memory cost is 19456 KiB, the parallelism is 1
and the output length is 32. To produce a hash with a random salt, you
can use the following command:

argon2 $(head -c 16 /dev/urandom | base64) -id -t 2 -k 19456 -p 1 -l 32 -e | tr -d '\n' > /data/passwd.argon2id